requirements.txt
faicons
shiny @ git+https://github.com/posit-dev/py-shiny
shinywidgets
plotly
pandas
ridgeplot
openai
python-dotenv
duckdb